requirements.txt
prophet
Django
django_plotly_dash
django_apscheduler
neomodel
apscheduler
scikit-learn
tensorflow
Cython
owlready2
dash_renderer
dpd_static_support
channels_redis